Added documentation in docs/upgradeRecipies/upgradeRecipe-2.0.0.adoc for the change to Parser.parse. Pulled SyntaxAnalysisMode into its own file and moved getSyntaxAnalysisModeFromInt into this file. Renamed ParserTest to SyntaxAnalysisModeTest to reflect this change.
